requirements.txt
streamlit
tensorflow
numpy
matplotlib
pillow
pdfplumber
ollama
whisper
yt_dlp
youtube_transcript_api
frontend
tools
pymupdf